Add a lock file option to ensure only one geoipupdate can run at once
This new option is called LockFile. It is a path to a file to lock. By
default it is not set and no locking will occur, so the default
behaviour is unchanged.

At program startup and prior to performing any updates, we now try to
acquire a lock on a file (if a lock file is set). If we can't, we abort
with an error.

This is to avoid race condition errors that have been seen in the past
where multiple geoipupdate processes ran concurrently.