Switch to waiting on a lock, and always locking
We now default to having a lock file set under the database directory.
The lock file can still be changed via the configuration file.

Note this changes the behaviour to always attempt to lock. There is no
option to disable locking. I'm thinking such an option may not be
necessary.