Report what went wrong with a download more accurately
Previously the download service generally responded with HTTP 200, even
in the case of several types of errors. The only way to tell that
something went wrong and that we did not download a database was to
inspect the body to see that it was a message rather than a gzip file.

There was only one such message we handled, "no new updates available".
Other cases reported only that the response was not a valid gzip file.
This was not informative.

The download service now responds with an HTTP status other than 200 for
most error cases, so we check for that and report what it said in these
cases.

We still handle the "no new updates available" specially. For backwards
compatibility reasons the service still sends HTTP 200 in this case.
However rather than inspecting the body for this message, we can now
check whether the `X-Database-MD5` contains a hash that matches the
database we already have.

For all other HTTP 200 responses, we require that the body is gzip.

The backwards compatibility issue is this: If the service sent something
like HTTP 304 when there were no updates available, then geoipupdate
clients prior to this update would say something like "received an
unexpected HTTP status code" but give no details.