Add environment variable configuration options
Previous various parts of the config could only be set in a config
file (and some parts by the config file or command line options).
This updates the config package to also set config options based
on environment variables after the config file and before command
line options are used.